wip: code cleanup & improved project list page
