+++ packages/nc-gui/components/projectList/emptyProjectListPlaceholder.vue
@@ -1,0 +1,17 @@
+<template>
+  <div
+    class="px-4 py-10 text-center textColor--text text--lighten-3 caption backgroundColor "
+  >
+    Please create a project
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'EmptyProjectListPlaceholder'
+}
+</script>
+
+<style scoped>
+
+</style>
